+++ src/main/java/me/despical/commonsbox/scoreboard/common/EntryBuilder.java
@@ -21,6 +21,7 @@ package me.despical.commonsbox.scoreboard.common;
 import java.util.LinkedList;
 import java.util.List;
 
+import me.despical.commonsbox.compat.VersionResolver;
 import me.despical.commonsbox.scoreboard.type.Entry;
 
 /**
@@ -42,13 +43,20 @@ public final class EntryBuilder {
 	}
 
 	public List<Entry> build() {
-		entries.forEach(entry -> entry.setPosition(entries.size() - entry.getPosition()));
+		for (Entry entry : entries) {
+			entry.setPosition(entries.size() - entry.getPosition());
+		}
+
 		return entries;
 	}
 
 	private String adapt(String entry) {
-		if (entry.length() > 48)
-			entry = entry.substring(0, 47);
+		if (VersionResolver.isCurrentEqualOrHigher(VersionResolver.ServerVersion.v1_14_R1)) {
+			if (entry.length() > 144) entry = entry.substring(0, 143);
+		} else {
+			if (entry.length() > 48) entry = entry.substring(0, 47);
+		}
+
 		return Strings.format(entry);
 	}
 }